job-info: Refactor module multiple files
Refactor code base into multiple files.
